chore: address post-merge housekeeping items
- Moved `FuzzNormalizeSingleQuoted` out of `paths_fuzz_test.go` and into `normalize_fuzz_test.go` to cleanly separate string normalization fuzzing from path boundary fuzzing.
- Increased audit log fidelity in `shellTool.ExecuteCommand` by logging both the original `COMMAND` and the final `EXECUTED` argv string whenever they differ (e.g., when wrapped in `sh -c` or translated for Windows).
