+++ internal/tools/workspace/shell.go
@@ -231,7 +231,12 @@ func (t *shellTool) ExecuteCommand(ctx context.Context, args map[string]interfac
 		return tools.ToolResult{Error: err, Text: fmt.Sprintf("Error: %v", err)}, nil
 	}
 
-	outputFile, approved, err := t.authorizeAndAudit(ctx, params, displayCmd)
+	// Build the actual executed command string for audit fidelity.
+	// When prepareCommand wraps the command (e.g. sh -c "..."), parts
+	// reflects what is actually executed.
+	executedCmdStr := strings.Join(parts, " ")
+
+	outputFile, approved, err := t.authorizeAndAudit(ctx, params, displayCmd, executedCmdStr)
 	if err != nil || !approved {
 		return t.handleAuthResult(approved, err, "command: "+displayCmd)
 	}
@@ -302,7 +307,7 @@ func (t *shellTool) prepareExecutionParts(params executeParams) ([]string, strin
 	return parts, params.Command, nil
 }
 
-func (t *shellTool) authorizeAndAudit(ctx context.Context, params executeParams, displayCommand string) (string, bool, error) {
+func (t *shellTool) authorizeAndAudit(ctx context.Context, params executeParams, displayCommand string, executedCommand string) (string, bool, error) {
 	outputFile, err := t.resolveOutputFile(params.OutputFile)
 	if err != nil {
 		return "", false, err
@@ -314,7 +319,7 @@ func (t *shellTool) authorizeAndAudit(ctx context.Context, params executeParams,
 		return outputFile, approved, err
 	}
 
-	t.auditExecution(displayCommand, params.Reason, params.OutputFile, params.Append)
+	t.auditExecution(displayCommand, executedCommand, params.Reason, params.OutputFile, params.Append)
 	return outputFile, true, nil
 }
 
@@ -571,10 +576,13 @@ func (t *shellTool) startHeartbeat(hb chan<- struct{}) (stop func()) {
 	}
 }
 
-func (t *shellTool) auditExecution(command, reason, outputFile string, isAppend bool) {
+func (t *shellTool) auditExecution(command, executedCommand, reason, outputFile string, isAppend bool) {
 	argsAudit := []any{
 		"REASON", reason,
 		"COMMAND", command,
+	}
+	if command != executedCommand {
+		argsAudit = append(argsAudit, "EXECUTED", executedCommand)
 	}
 	if outputFile != "" {
 		argsAudit = append(argsAudit, "OUTPUT_FILE", outputFile, "APPEND", isAppend)
